requirements.txt
selenium==3.141.0
python-telegram-bot==13.6
beautifulsoup4==4.9.0
